WIP: path functions

+++ inc/path.h
@@ -15,41 +15,41 @@ namespace rtm {
 	const char* pathGetFileName(const char* _path);
 
     /// Retrieves file name with extension if it is a file path
-	bool pathGetFilename(const char* _path, char* _buffer, size_t _bufferSize);
+	bool pathGetFilename(const char* _path, char* _buffer, uint32_t _bufferSize);
 
 	/// Retrieves file name without extension if it is a file path
-	bool pathGetFilenameNoExt(const char* _path, char* _buffer, size_t _bufferSize);
+	bool pathGetFilenameNoExt(const char* _path, char* _buffer, uint32_t _bufferSize);
 
 	/// Retrieves file extension if it is a file path
 	const char* pathGetExt(const char* _path);
 
 	/// Retrieves file extension if it is a file path
-	bool pathGetExt(const char* _path, char* _buffer, size_t _bufferSize);
+	bool pathGetExt(const char* _path, char* _buffer, uint32_t _bufferSize);
 
 	/// Retrieves current working directory
-	bool pathGetCurrentDirectory(char* _buffer, size_t _bufferSize);
+	bool pathGetCurrentDirectory(char* _buffer, uint32_t _bufferSize);
 
 	/// Retrieves path to data root directory
-	bool pathGetDataDirectory(char* _buffer, size_t _bufferSize);
+	bool pathGetDataDirectory(char* _buffer, uint32_t _bufferSize);
 
 	/// Appends two paths and stores result to path referred by _path handle
-	bool pathAppend(const char* _path, const char* _appendPath, char* _buffer, size_t _bufferSize);
+	bool pathAppend(const char* _path, const char* _appendPath, char* _buffer, uint32_t _bufferSize);
 
 	/// Goes 'up' one level 
-	bool pathUp(const char* _path, char* _buffer, size_t _bufferSize);
+	bool pathUp(const char* _path, char* _buffer, uint32_t _bufferSize);
 
 	/// Removes relative portions of the path
-	bool pathCanonicalize(const char* _path, char* _buffer, size_t _bufferSize);
+	bool pathCanonicalize(const char* _path, char* _buffer, uint32_t _bufferSize);
 
     /// Removes relative portions of the path In place 
     void pathCanonicalize(char* _path);
 
 	/// Makes _relative path absolute by appending it to _base
 	/// and repeats using current directory if path is still relative
-	bool pathMakeAbsolute(const char* _relative, const char* _base, char* _buffer, size_t _bufferSize);
+	bool pathMakeAbsolute(const char* _relative, const char* _base, char* _buffer, uint32_t _bufferSize);
 
 	/// Makes _path relative
-	bool pathMakeRelative(const char* _pathFrom, bool _fromDir, const char* _pathTo, bool _toDir, char* _buffer, size_t _bufferSize);
+	bool pathMakeRelative(const char* _pathFrom, bool _fromDir, const char* _pathTo, bool _toDir, char* _buffer, uint32_t _bufferSize);
 	
 	/// Checks if path is absolute
 	bool pathIsAbsolute(const char* _path);
